Bug Fixes
Fixed users ballance being a negative after item purchase
Fixed invalid command handling for
- moneyInput()
- Purchase confirmation
Added display of users ballance during item selection
Added check to see if balance is empty and display differnet message
varying on if the balance is 0 or not.
Added check to set users balance to 0 if it somehow is a negetive value.
Added wait timer to exit().